+++ src/index.html
@@ -5,12 +5,12 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mesto</title>
-  <link rel="stylesheet" href="./index.css">
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
 </head>
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('./images/logo.svg')%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -32,13 +32,13 @@
 	<div class="popup-image">
 		<div class="popup-image__block">
 			<img src="" class="popup-image__content">
-			<img src="./images/close.svg" alt="" class="popup-image__close">
+			<img src="<%=require('./images/close.svg')%>" alt="" class="popup-image__close">
 		</div>	
 	</div>	
    
     <div class="popup">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Новое место</h3>
         <form class="popup__form" name="newCard">
             <input type="text" onkeyup="checkInputPopup()" name="name" class="popup__input popup__input_type_name" placeholder="Название">
@@ -50,7 +50,7 @@
     
     <div class="popup-profile">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup-profile__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup-profile__close">
         <h3 class="popup__title">Редактировать профиль</h3>
         <form class="popup__form" name="newProfile">
             <input type="text" onkeyup="checkInputPopup()" name="profilename" class="popup__input popup__input_type_name-prof" placeholder="Имя" value="Plamenevskiy Nikita" autocomplete="off" minlength="2" maxlength="30" required/>        
@@ -63,18 +63,19 @@
     </div>
     
   </div>
-  
-  <script src='./modules/api.js'></script>
-  <script src='./modules/checkValidity.js'></script>
-  <script src='./modules/checkInputPopup.js'></script>
-  <script src='./modules/addUserInfo.js'></script>
-  <script src='./modules/Popup.js'></script>
-  <script src='./modules/initialCards.js'></script>
-  <script src='./modules/Card.js'></script>
-  <script src='./modules/CardList.js'></script>
-  <script src="./modules/script.js"></script>
-  <script src="./index.js"></script>
-  
+    <script src='/src/modules/addUserInfo.js'></script> 
+    <script src='/src/modules/Popup.js'></script> 
+    <script src="/src/modules/CardList.js"></script> 
+    <script src='/src/modules/checkValidity.js'></script> 
+    <script type='module' src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
+    <script src='/src/modules/script.js'></script> 
+    <script src='/src/modules/api.js'></script> 
+    <script src="/src/modules/takeUserInfo.js"></script> 
+    <script src='/src/modules/checkInputPopup.js'></script> 
+    <script src='/src/modules/Card.js'></script> 
+    <script src='/src/modules/initialCards.js'></script> 
+    
+    
 </body>
 </html>
