+++ src/modules/CardList.js
@@ -16,5 +16,3 @@ class CardList {
     });
   }
 }
-
-
