+++ src/modules/checkValidity.js
@@ -3,10 +3,7 @@ function checkValidity(event) {
   const about = profileForm.elements.about;
   const errorName = document.querySelector('.popup__error_name');
   const errorAbout = document.querySelector('.popup__error_about');
-  // Можно лучше: обычно названия, для примера 'Должно быть от 2 до 30 символов' 
-  // выносят в отдельный объект. Допустим может появится задача сделать многоязычный сайт
-  // Для примера : const words = { validationLenght: 'Должно быть от 2 до 30 символов'	} 
-  // Далее words передаётся в функцию и используется.
+
   if (profilename.value.length === 0) {
     errorName.textContent = 'Это обязательное поле';
   } else if (profilename.value.length < 2) {
@@ -22,5 +19,4 @@ function checkValidity(event) {
   } else {
     errorAbout.textContent = null;
   }
-
 }
